Remove: All modules in favour of single module
This new module is a rewrite of the loopjs logic that provides much
of the same functionality as all the previous modules. This change
was introduced as an opportunity to revisit the iterator logic and
increase the amount of code reuse between loops.

Overall, this has appeared successful. Some features were dropped
because they were unused. There is no technical limitation preventing
the removed features from being added again.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>